Use juju-2.1 as a fallback. (#39)
Currently "juju-2.0" is hard-coded as the binary name. This change adds more flexibility and makes an attempt at using a binary that actually exists.

Note that this patch does more than the minimum to fix the bug. This is because we will likely run into this problem again and again, so a slightly broader solution is warranted. At the same time, an even broader solution is arguably valid here by the same reasoning. However, that can be addressed separately if it's worth bothering.

Also note that this branch does not actually make any effort toward compatibility with Juju 2.1 (if any is required). Instead it opens the door for such efforts, allowing us to run the landscape system tests using 2.1.