Re-enable integration tests
Integration tests exercising txjuju with fake-juju where disabled at
some point in the past, probably during the migration to Juju 2 and/or
to GitHub/Travis.

This branch restores them, with the necessary modifications to run
against fake-juju 2.0.2.

Since Juju 2 has no direct equivalent of the "model info" API, that
one test was dropped.